+++ lib/src/normalizer/normalize.dart
@@ -1,0 +1,6 @@
+import 'package:ml_linalg/norm.dart';
+import 'package:ml_preprocessing/src/normalizer/normalizer.dart';
+import 'package:ml_preprocessing/src/pipeline/pipeable.dart';
+
+PipeableOperatorFn normalize([Norm norm = Norm.euclidean]) =>
+        (_) => Normalizer(norm);